// spark-event-ui/src/app/rest/model/sparkAppMetricsDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TaskMetrics } from './taskMetrics';


export interface SparkAppMetricsDTO { 
    invalidEvent?: number;
    logStart?: number;
    applicationStart?: number;
    applicationEnd?: number;
    resourceProfileAdded?: number;
    environmentUpdate?: number;
    blockManagerRemoved?: number;
    blockManagerAdded?: number;
    executorAdded?: number;
    executorRemoved?: number;
    executorExcluded?: number;
    executorExcludedForStage?: number;
    nodeExcludedForStage?: number;
    executorUnexcluded?: number;
    nodeExcluded?: number;
    nodeUnexcluded?: number;
    executorMetricsUpdate?: number;
    sqlJobStart?: number;
    topLevelJobStart?: number;
    sqlJobEndSucceed?: number;
    sqlJobEndFailed?: number;
    topLevelJobEndSucceed?: number;
    topLevelJobEndFailed?: number;
    stageSubmitted?: number;
    stageCompleted?: number;
    stageCompletedSkip?: number;
    stageCompletedFailed?: number;
    stageCompletedPending?: number;
    stageExecutorMetrics?: number;
    taskStart?: number;
    speculativeTaskSubmitted?: number;
    taskGettingResult?: number;
    unschedulableTaskSetAdded?: number;
    unschedulableTaskSetRemoved?: number;
    taskEnd?: number;
    unpersistRDD?: number;
    blockUpdated?: number;
    sqlExecutionStart?: number;
    sqlExecutionEnd?: number;
    sqlAdaptiveExecutionUpdate?: number;
    sqlAdaptiveMetricUpdates?: number;
    sqlDriverAccumUpdates?: number;
    miscellaneousProcessAdded?: number;
    otherEvent?: number;
    totalTaskMetrics?: TaskMetrics;
    spillMemoryTaskCount?: number;
    spillDiskTaskCount?: number;
}